requirements.txt
# ============================================
# SAC-Diffusion Driving Policy Requirements
# ============================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
einops>=0.6.1

# Reinforcement Learning
gymnasium>=0.28.1
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Isaac Lab / Isaac Sim (Install separately from NVIDIA)
# See: https://isaac-sim.github.io/IsaacLab/
# isaacsim>=4.0.0  # Install via NVIDIA instructions

# Robotics
numpy>=1.24.0
scipy>=1.10.0
transforms3d>=0.4.1
pyquaternion>=0.9.9

# ROS2 Interface (for real robot deployment)
# rclpy  # Install via ROS2 Humble apt packages

# Point Cloud Processing
open3d>=0.17.0
torch-geometric>=2.3.0  # For PointNet-like encoders

# Vision
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Experiment Management & Logging
wandb>=0.15.0
tensorboard>=2.13.0
tensorboardX>=2.6.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# Data Processing
h5py>=3.9.0
zarr>=2.15.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
rich>=13.4.0
termcolor>=2.3.0

# Code Quality
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Diffusion Model Specific
diffusers>=0.18.0  # Optional: HuggingFace diffusers for reference
accelerate>=0.20.0

# Math & Scientific Computing
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Profiling & Performance
py-spy>=0.3.14
memory_profiler>=0.61.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# 3D Visualization (optional)
pyvista>=0.41.0
trimesh>=3.23.0

# Safety & Monitoring
safetensors>=0.3.1

# Hyperparameter Optimization (optional)
optuna>=3.2.0
ray[tune]>=2.6.0

# Multi-processing
joblib>=1.3.0
cloudpickle>=2.2.0
